move zoekt.FakeSearcher to backend.FakeSearcher
