+++ cmd/frontend/graphqlbackend/search_repositories_test.go
@@ -12,7 +12,6 @@ import (
 	"github.com/google/go-cmp/cmp"
 	"github.com/google/go-cmp/cmp/cmpopts"
 
-	searchzoekt "github.com/sourcegraph/sourcegraph/cmd/frontend/internal/search/zoekt"
 	"github.com/sourcegraph/sourcegraph/internal/api"
 	"github.com/sourcegraph/sourcegraph/internal/search"
 	searchbackend "github.com/sourcegraph/sourcegraph/internal/search/backend"
@@ -27,7 +26,7 @@ func TestSearchRepositories(t *testing.T) {
 		{Repo: &types.RepoName{ID: 789, Name: "bar/one"}, Revs: []search.RevisionSpecifier{{RevSpec: ""}}},
 	}
 
-	zoekt := &searchbackend.Zoekt{Client: &searchzoekt.FakeSearcher{}}
+	zoekt := &searchbackend.Zoekt{Client: &searchbackend.FakeSearcher{}}
 
 	mockSearchFilesInRepos = func(args *search.TextParameters) (matches []*FileMatchResolver, common *SearchResultsCommon, err error) {
 		repos, err := getRepos(context.Background(), args.RepoPromise)
@@ -148,7 +147,7 @@ func TestRepoShouldBeAdded(t *testing.T) {
 		}
 	}
 
-	zoekt := &searchbackend.Zoekt{Client: &searchzoekt.FakeSearcher{}}
+	zoekt := &searchbackend.Zoekt{Client: &searchbackend.FakeSearcher{}}
 
 	t.Run("repo should be included in results, query has repoHasFile filter", func(t *testing.T) {
 		repo := &search.RepositoryRevisions{Repo: &types.RepoName{ID: 123, Name: "foo/one"}, Revs: []search.RevisionSpecifier{{RevSpec: ""}}}
